remove useServerUrls, move checker under ui
