+++ cloudbuild.yaml
@@ -1,36 +1,35 @@
 steps:
   # 構建 Docker 映像
-  - name: 'gcr.io/cloud-builders/docker'
-    args: ['build', '-t', 'gcr.io/$PROJECT_ID/agent-chat-ui', '.']
-  
+  - name: "gcr.io/cloud-builders/docker"
+    args: ["build", "-t", "gcr.io/$PROJECT_ID/agent-chat-ui", "."]
+
   # 推送映像到 Container Registry
-  - name: 'gcr.io/cloud-builders/docker'
-    args: ['push', 'gcr.io/$PROJECT_ID/agent-chat-ui']
-  
+  - name: "gcr.io/cloud-builders/docker"
+    args: ["push", "gcr.io/$PROJECT_ID/agent-chat-ui"]
+
   # 部署到 Cloud Run
-  - name: 'gcr.io/google.com/cloudsdktool/cloud-sdk'
+  - name: "gcr.io/google.com/cloudsdktool/cloud-sdk"
     entrypoint: gcloud
     args:
-      - 'run'
-      - 'deploy'
-      - 'agent-chat-ui'
-      - '--image'
-      - 'gcr.io/$PROJECT_ID/agent-chat-ui'
-      - '--platform'
-      - 'managed'
-      - '--region'
-      - 'asia-east1'
-      - '--allow-unauthenticated'
-      - '--set-env-vars'
-      - 'NEXT_PUBLIC_API_URL=${_NEXT_PUBLIC_API_URL},NEXT_PUBLIC_ASSISTANT_ID=${_NEXT_PUBLIC_ASSISTANT_ID},GSTUDIO_API=${_GSTUDIO_API}'
+      - "run"
+      - "deploy"
+      - "agent-chat-ui"
+      - "--image"
+      - "gcr.io/$PROJECT_ID/agent-chat-ui"
+      - "--platform"
+      - "managed"
+      - "--region"
+      - "asia-east1"
+      - "--allow-unauthenticated"
+      - "--set-env-vars"
+      - "NEXT_PUBLIC_API_URL=${_NEXT_PUBLIC_API_URL},NEXT_PUBLIC_ASSISTANT_ID=${_NEXT_PUBLIC_ASSISTANT_ID},GSTUDIO_API=${_GSTUDIO_API}"
 
 images:
-  - 'gcr.io/$PROJECT_ID/agent-chat-ui'
+  - "gcr.io/$PROJECT_ID/agent-chat-ui"
 
 substitutions:
-  _NEXT_PUBLIC_API_URL: 'https://your-langgraph-api.run.app'
-  _NEXT_PUBLIC_ASSISTANT_ID: 'agent'
-  _GSTUDIO_API: 'https://your-gstudio-api.run.app'
-
+  _NEXT_PUBLIC_API_URL: "https://your-langgraph-api.run.app"
+  _NEXT_PUBLIC_ASSISTANT_ID: "agent"
+  _GSTUDIO_API: "https://your-gstudio-api.run.app"
 # 您可以在觸發構建時覆蓋這些默認值
 # 例如: gcloud builds submit --config=cloudbuild.yaml --substitutions=_NEXT_PUBLIC_API_URL="https://new-api.run.app",_GSTUDIO_API="https://new-gstudio.run.app"
